Agregar policy de vacunas

+++ mascotas-api/app/Http/Controllers/VaccineController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\VaccineRequest;
+use App\Repositories\PetRepository;
 use App\Repositories\VaccineRepository;
 use Illuminate\Http\JsonResponse;
 
@@ -10,9 +11,10 @@ class VaccineController extends Controller
 {
     protected $vaccineRepository;
 
-    public function __construct(VaccineRepository $vaccineRepository)
+    public function __construct(VaccineRepository $vaccineRepository, PetRepository $petRepository)
     {
         $this->vaccineRepository = $vaccineRepository;
+        $this->petRepository = $petRepository;
     }
 
     /**
@@ -23,6 +25,9 @@ class VaccineController extends Controller
      */
     public function getVaccinesByPet($pet_id): JsonResponse
     {
+        $pet = $this->petRepository->find($pet_id);
+        $this->authorize('viewMedicalDetails', $pet);
+
         $vaccines = $this->vaccineRepository->getVaccinesByPet($pet_id)->values();
         return response()->json([
             'data' => compact('vaccines'),
@@ -54,6 +59,7 @@ class VaccineController extends Controller
     public function findVaccine($vaccine_id): JsonResponse
     {
         $vaccine = $this->vaccineRepository->find($vaccine_id);
+        $this->authorize('view', $vaccine);
 
         return response()->json([
             'data' => compact('vaccine'),
@@ -69,6 +75,9 @@ class VaccineController extends Controller
      */
     public function updateVaccine(VaccineRequest $request, $vaccine_id): JsonResponse
     {
+        $vaccine = $this->vaccineRepository->find($vaccine_id);
+        $this->authorize('update', $vaccine);
+
         $vaccine = $this->vaccineRepository->update($vaccine_id, $request->get('name'), $request->get('date'));
 
         return response()->json([
@@ -85,6 +94,9 @@ class VaccineController extends Controller
      */
     public function deleteVaccine($vaccine_id): JsonResponse
     {
+        $vaccine = $this->vaccineRepository->find($vaccine_id);
+        $this->authorize('delete', $vaccine);
+
         $this->vaccineRepository->delete($vaccine_id);
 
         return response()->json([
